+++ tests/test_local_search_quantizer.py
@@ -572,3 +572,48 @@ class TestProductLocalSearchQuantizer(unittest.TestCase):
 
         # max rtoal in OSX: 2.87e-6
         np.testing.assert_allclose(lut, lut_ref, rtol=5e-06)
+
+
+class TestIndexProductLocalSearchQuantizer(unittest.TestCase):
+
+    def test_accuracy1(self):
+        """check that the error is in the same ballpark as LSQ."""
+        recall1 = self.eval_index_accuracy("PLSQ4x3x5_Nqint8")
+        recall2 = self.eval_index_accuracy("LSQ12x5_Nqint8")
+        self.assertGreaterEqual(recall1, recall2)  # 622 vs 551
+
+    def test_accuracy2(self):
+        """when nsplits = 1, PLSQ should be the same as LSQ"""
+        recall1 = self.eval_index_accuracy("PLSQ1x3x5_Nqint8")
+        recall2 = self.eval_index_accuracy("LSQ3x5_Nqint8")
+        self.assertEqual(recall1, recall2)
+
+    def eval_index_accuracy(self, index_key):
+        ds = datasets.SyntheticDataset(32, 1000, 1000, 100)
+        index = faiss.index_factory(ds.d, index_key)
+
+        index.train(ds.get_train())
+        index.add(ds.get_database())
+        D, I = index.search(ds.get_queries(), 10)
+        inter = faiss.eval_intersection(I, ds.get_groundtruth(10))
+
+        # do a little I/O test
+        index2 = faiss.deserialize_index(faiss.serialize_index(index))
+        D2, I2 = index2.search(ds.get_queries(), 10)
+        np.testing.assert_array_equal(I2, I)
+        np.testing.assert_array_equal(D2, D)
+
+        return inter
+
+    def test_factory(self):
+        AQ = faiss.AdditiveQuantizer
+        ns, Msub, nbits = 2, 4, 8
+        index = faiss.index_factory(64, f"PLSQ{ns}x{Msub}x{nbits}_Nqint8")
+        assert isinstance(index, faiss.IndexProductLocalSearchQuantizer)
+        self.assertEqual(index.plsq.nsplits, ns)
+        self.assertEqual(index.plsq.subquantizer(0).M, Msub)
+        self.assertEqual(index.plsq.subquantizer(0).nbits.at(0), nbits)
+        self.assertEqual(index.plsq.search_type, AQ.ST_norm_qint8)
+
+        code_size = (ns * Msub * nbits + 7) // 8 + 1
+        self.assertEqual(index.plsq.code_size, code_size)