+++ tests/test_residual_quantizer.py
@@ -1176,3 +1176,48 @@ class TestProductResidualQuantizer(unittest.TestCase):
 
         print(err_prq, err_rq)
         self.assertEqual(err_prq, err_rq)
+
+
+class TestIndexProductResidualQuantizer(unittest.TestCase):
+
+    def test_accuracy1(self):
+        """check that the error is in the same ballpark as RQ."""
+        recall1 = self.eval_index_accuracy("PRQ4x3x5_Nqint8")
+        recall2 = self.eval_index_accuracy("RQ12x5_Nqint8")
+        self.assertGreaterEqual(recall1 * 1.1, recall2)  # 657 vs 665
+
+    def test_accuracy2(self):
+        """when nsplits = 1, PRQ should be the same as RQ"""
+        recall1 = self.eval_index_accuracy("PRQ1x3x5_Nqint8")
+        recall2 = self.eval_index_accuracy("RQ3x5_Nqint8")
+        self.assertEqual(recall1, recall2)
+
+    def eval_index_accuracy(self, index_key):
+        ds = datasets.SyntheticDataset(32, 1000, 1000, 100)
+        index = faiss.index_factory(ds.d, index_key)
+
+        index.train(ds.get_train())
+        index.add(ds.get_database())
+        D, I = index.search(ds.get_queries(), 10)
+        inter = faiss.eval_intersection(I, ds.get_groundtruth(10))
+
+        # do a little I/O test
+        index2 = faiss.deserialize_index(faiss.serialize_index(index))
+        D2, I2 = index2.search(ds.get_queries(), 10)
+        np.testing.assert_array_equal(I2, I)
+        np.testing.assert_array_equal(D2, D)
+
+        return inter
+
+    def test_factory(self):
+        AQ = faiss.AdditiveQuantizer
+        ns, Msub, nbits = 2, 4, 8
+        index = faiss.index_factory(64, f"PRQ{ns}x{Msub}x{nbits}_Nqint8")
+        assert isinstance(index, faiss.IndexProductResidualQuantizer)
+        self.assertEqual(index.prq.nsplits, ns)
+        self.assertEqual(index.prq.subquantizer(0).M, Msub)
+        self.assertEqual(index.prq.subquantizer(0).nbits.at(0), nbits)
+        self.assertEqual(index.prq.search_type, AQ.ST_norm_qint8)
+
+        code_size = (ns * Msub * nbits + 7) // 8 + 1
+        self.assertEqual(index.prq.code_size, code_size)